+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to the
 [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
+## 0.1.2.0 - 2025-05-13
+### Updated
+Fixed the newStruct logic
+implemented pCommaParam into our trait,impl,func,method so we can have multiple params
+Code generator work
+
 ## 0.1.0.0 - 2025-03-22
 
 ### Added